# backend/docs/ai-converse.openapi.yaml
openapi: 3.0.3
info:
  title: ExpenseBeam AI — Conversational Expense Logging, Editing & Q&A
  description: >
    Spec for `POST /api/ai/converse`, plus the two existing `/api/ai`
    endpoints it shares auth/quota/caching behavior with, for context. As of
    1.1.0, converse also resolves deletion requests via `delete_intent`,
    paired with `DELETE /api/expenses/batch`. As of 1.2.0 it is the single
    entry point for the whole assistant, adding: `answering` (a finance
    question, resolved server-side against the user's real data via the same
    context as `/ai/ask`), `edit_intent`, `income_concluded`,
    `recurring_concluded` and `report_intent`.

    Stateless: the client resends the full message history each turn; there
    is no server-side session. For anything that writes or deletes, this
    endpoint resolves *intent* and a *filter* only — never a count and never
    the affected rows. The client independently re-verifies the exact
    matching transactions via `GET /expenses` and confirms with the user
    before deleting or editing anything.

    Client-gated: a client sends `supports` listing the statuses it can carry
    through to completion. Anything outside that list is downgraded on the way
    out — `answering` becomes a plain `asking` turn carrying the same prose,
    and the write intents become a short message pointing at the equivalent
    screen. A request with no `supports` is treated as an older build and
...

...
    conversation itself keeps changing.
...
servers:
...
security:
  - bearerAuth: []
paths:
  /ai/converse:
    post:
      summary: Conversational expense logging / deletion turn
      description: >
        Send the full conversation so far (the client's own static opening
        greeting is NOT included — `messages` starts at the first real user
        turn). Returns one of three states: still gathering info for a new
        expense (`asking`), a finished expense list ready to save
        (`concluded`), or a resolved deletion request (`delete_intent`).
        Counts against the same 100/month AI quota as `/ai/parse` and
        `/ai/ask`, with the same 24h response cache (keyed on the full
        serialized message history, so replaying an identical conversation
        returns a cached hit and does not consume quota).


        **Safety-critical prompting note for `delete_intent`**: the backend
        classifies a turn as `delete_intent` ONLY on unambiguous deletion
        language ("delete", "clear", "remove", "get rid of" + a scope). A
        message like "what did I spend this month" is a read query, not a
        deletion request, and is never classified as `delete_intent`. When
        intent is ambiguous, the response is `asking` with a clarifying
        question instead of guessing — this endpoint drives an irreversible
        action downstream, so under-triggering is far safer than
        over-triggering. As defense in depth, `filter.category` is always
        re-validated server-side against the user's real category list
        (case-insensitive match, normalized to the stored name) and
        `filter.start_date`/`filter.end_date` are validated as `YYYY-MM-DD`
        — anything that doesn't match is nulled out rather than passed
        through from the model's raw output.
      operationId: converseExpenses
      tags: [AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConverseRequest'
            examples:
              firstTurn:
                summary: First user message
                value:
                  messages:
                    - role: user
                      content: coffee and lunch today
              followUp:
                summary: Mid-conversation, answering a clarifying question
                value:
                  messages:
                    - role: user
                      content: coffee and lunch today
                    - role: assistant
                      content: Got it — how much was the coffee?
                    - role: user
                      content: 45 EGP, lunch was 320
              deleteRequest:
                summary: Deletion request
                value:
                  messages:
                    - role: user
                      content: clear this month's transactions
              deleteRequestWithCategory:
                summary: Deletion request scoped by date + category
                value:
                  messages:
                    - role: user
                      content: delete my coffee expenses last week
      responses:
        '200':
          description: >
            Discriminated by `status`. See `ConverseResponse` for the full
            shape, including the `delete_intent` branch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConverseResponse'
              examples:
                asking:
                  summary: Still gathering info
                  value:
                    status: asking
                    message: What category for the lunch?
                    quick_replies: ["Food & Dining", "Other"]
                concluded:
                  summary: Expense logging done — matches the shape ConfirmExpensesScreen already parses
                  value:
                    status: concluded
                    message: "Here's what I found:"
                    expenses:
                      - description: Coffee
                        amount: 45
                        currency: EGP
                        category: Food & Dining
                        date: "2026-09-20T00:00:00.000Z"
                      - description: Lunch
                        amount: 320
                        currency: EGP
                        category: Food & Dining
                        date: "2026-09-20T00:00:00.000Z"
                deleteIntentDateRange:
                  summary: Deletion request resolved to a date range
                  value:
                    status: delete_intent
                    message: You want to clear all transactions from September 2026?
                    filter:
                      start_date: "2026-09-01"
                      end_date: "2026-09-30"
                      category: null
                deleteIntentWithCategory:
                  summary: Deletion request resolved to a date range + category
                  value:
                    status: delete_intent
                    message: Delete your Food & Dining expenses from the last 7 days?
                    filter:
                      start_date: "2026-09-13"
                      end_date: "2026-09-20"
                      category: Food & Dining
                cachedHit:
                  summary: Identical conversation replayed within 24h (does not consume quota)
                  value:
                    status: asking
                    message: What category for the lunch?
                    quick_replies: ["Food & Dining", "Other"]
                    cached: true
        '400':
          description: Missing or malformed `messages`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing:
                  value: { error: "messages is required" }
                malformed:
                  value: { error: "Invalid messages format" }
        '401':
          description: Missing/invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { error: "No token provided" }
        '429':
          description: Monthly AI quota exceeded (100 calls/month, shared with `/parse` and `/ask`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotaExceededResponse'
              example:
                error: quota_exceeded
                message: Monthly AI limit reached.
                reset_date: "2026-10-01"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ChatTurn:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
          minLength: 1
    ConverseRequest:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ChatTurn'
    ParsedExpense:
      type: object
      description: >
        Shape consumed directly by the mobile client's
        ConfirmExpensesScreen (`_ExpenseForm.fromParsed`) — the same shape
        `/ai/parse` already returns.
      required: [description, amount, currency, category, date]
      properties:
        description: { type: string, maxLength: 60 }
        amount: { type: number }
        currency: { type: string, example: EGP }
        category: { type: string }
        date:
          type: string
          format: date-time
          description: ISO 8601. Defaults to today when not stated by the user.
    DeleteFilter:
      type: object
      description: >
        Resolved deletion scope. The client maps these directly onto the
        existing `GET /expenses` query params (`startDate`/`endDate`/
        `category`) it already uses elsewhere in the app — no new
        client-side filter vocabulary. `category`, when present, is always
        one of the user's real category names — the backend re-validates it
        against the caller's actual category list and nulls it out rather
        than passing through anything unrecognized.
      properties:
        start_date:
          type: string
          format: date
          nullable: true
        end_date:
          type: string
          format: date
          nullable: true
        category:
          type: string
          nullable: true
    ConverseResponse:
      type: object
      description: Discriminated union on `status`.
      required: [status, message]
      properties:
        status:
          type: string
          enum:
            - asking
            - concluded
            - delete_intent
            - answering
            - edit_intent
            - income_concluded
            - recurring_concluded
            - report_intent
          description: >
            `question_intent` is an internal status the model may return; it
            never reaches the client, because the route resolves it into
            `answering` within the same request.
        message:
          type: string
          description: >
            Clarifying question (asking), short friendly summary (concluded,
            income_concluded, recurring_concluded), the finished answer to a
            finance question (answering), or a confirmation-style summary of
            a resolved scope (delete_intent, edit_intent) — e.g. "You want to
            clear all transactions from September 2026?". For the intent
            statuses the client shows this as a normal chat bubble; it is NOT
            itself the confirmation step — the client always independently
            resolves the real matching rows and shows its own Cancel/Confirm
            card afterward.
        quick_replies:
          type: array
          items: { type: string }
          description: Only present/non-empty when status=asking. Tapping one sends it as the next user turn.
        expenses:
          type: array
          items:
            $ref: '#/components/schemas/ParsedExpense'
          description: Only present when status=concluded.
        filter:
          $ref: '#/components/schemas/DeleteFilter'
          description: >
            Present when status=delete_intent, and when status=edit_intent
            with an additional nullable `match` field (a short description
            substring, passed to GET /expenses as `search`). Carries no count
            — counts must always come from the client's own GET /expenses
            query, never be asserted by this endpoint. At least one filter
            field is guaranteed non-null: a filter that resolved to nothing
            would select the user's entire history, so the server downgrades
            it to `asking` instead.
        changes:
          type: object
          description: >
            Only present when status=edit_intent. At least one key, each
            server-validated: amount (positive number), description
            (non-empty), category (an existing category name), date
            (YYYY-MM-DD). Sent as-is as the PUT /expenses/{id} body.
          properties:
            amount: { type: number }
            description: { type: string }
            category: { type: string }
            date: { type: string, format: date }
        incomes:
          type: array
          description: >
            Only present when status=income_concluded. Each entry has amount
            (positive), currency, source (one of Salary, Business, Freelance,
            Investment, Rental, Gift, Other — anything else is coerced to
            Other), optional description, and date (defaults to today).
            Posted by the client to POST /income after confirmation.
          items:
            type: object
            properties:
              amount: { type: number }
              currency: { type: string }
              source: { type: string }
              description: { type: string, nullable: true }
              date: { type: string, format: date }
        recurring:
          type: object
          description: >
            Only present when status=recurring_concluded. Posted by the
            client to POST /recurring after confirmation — note that endpoint
            takes camelCase `nextDueDate`, so the client maps it.
          properties:
            amount: { type: number }
            currency: { type: string }
            category: { type: string, nullable: true }
            description: { type: string, nullable: true }
            interval: { type: string, enum: [daily, weekly, monthly] }
            next_due_date: { type: string, format: date }
        period:
          type: string
          enum: [today, yesterday, week, month]
          description: >
            Only present when status=report_intent. These are the only
            periods GET /reports/report.png accepts; a request for any other
            range is downgraded to `asking` rather than snapped to a
            near-miss. The client fetches the PNG itself and renders it
            inline in the conversation.
        cached:
          type: boolean
          description: >
            True when this exact conversation hit the 24h response cache.
            Never set for status=answering — answers are not cached.
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
    QuotaExceededResponse:
      type: object
      properties:
        error: { type: string, example: quota_exceeded }
        message: { type: string }
        reset_date: { type: string, format: date, example: "2026-10-01" }